requirements.txt
flask>=3.0.0
requests>=2.31.0
redis>=5.0.0
tenacity>=8.2.3
cachetools>=5.3.1
uhashring>=2.3
python-dotenv>=1.0.1
rdflib>=7.0.0
